pyproject.toml
[tool.poetry]
name = "declaraciones-feuc"
version = "0.1.0"
description = ""
authors = ["Agustín Covarrubias <agucova@uc.cl>"]

[tool.poetry.dependencies]
python = "3.8.3"
python-dotenv = "^0.13.0"
peewee = "^3.13.3"
oauthlib = "^3.1.0"
flask-login = "^0.5.0"
requests = "^2.23.0"
pyopenssl = "^19.1.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.8.3"
pytest-cov = "^2.9.0"
pytest-flask = "^1.0.0"
pytest-xdist = "^1.32.0"
pytest-clarity = "^0.3.0-alpha.0"
pytest-emoji = "^0.2.0"
rope = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
